implementacion de ktor , kotlin serialization y napier. dudo fuertemente que la estructura y la fomra del http client este correcta
